requirements.txt
google-api-core
google-auth
requests
